+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 3.0.2
+- `s_packages` package dependency upgraded
+
 ## 3.0.1
 - dependencies updated